moved lib into app
